requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.7
django-countries==7.2.1
django-embed-video==1.4.0
idna==3.3
Pillow==8.4.0
python-dotenv==0.19.2
pytz==2021.3
pyuca==1.2
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
